refactor: move utils.hypixel package to hypixel package
