+++ src/main/kotlin/net/llvg/exec/hypixel/skyblock/catacombs/map/scan/RoomEntry.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2025 Water-OR
+ * Copyright (C) 2025-2025 Water-OR
  *
  * This file is part of ExeClient
  *
@@ -17,7 +17,12 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.llvg.exec.utils.hypixel.skyblock.catacombs.map.scan
+package net.llvg.exec.hypixel.skyblock.catacombs.map.scan
 
-@Suppress("UNUSED_PARAMETER")
-class CatacombsScanFailedException(e: Throwable) : RuntimeException()
+import net.llvg.exec.utils.vector.Vec2I
+
+class RoomEntry(
+        val pos: Vec2I
+) {
+        var info: RoomInfo? = null
+}
